Fix buildlib when called as a subprocess
The argument order differs for the cdeps library vs. the individual
components. Thus, we need a different function to parse the command line
for the library. The new parse_command_line function is a direct copy of
the same function from buildlib.csm_share.

To handle this difference in command-line arguments in the two
situations, I have renamed the old file to buildlib_comps (which is
identical to the old buildlib); the individual data model components now
point to that file, leaving buildlib just being used for the cdeps
shared library.

In a following commit I will remove unnecessary code from each of these
files.

Resolves ESCOMP/CDEPS#104