fix for zero pressure in nldas2 configuration (#159)
The problem encountered in nldas2 is that both the mesh had a mask of zero over ocean points and the route handle for the mapping of streams to the model resolution uses zeroregion=ESMF_REGION_TOTAL. As a result, all mapped values over ocean were set to zero. This is a real problem if the variable is surface pressure. Since the values over ocean are never used, the solution is to simply set a valid value of surface pressure over ocean.